+++ Arrangement_on_surface_2/test/Arrangement_on_surface_2/Point_location_test.h
@@ -33,7 +33,7 @@ public:
   typedef typename Base::Points_vector                  Points_vector;
   typedef typename Base::Xcurves_vector                 Xcurves_vector;
   typedef typename Base::Curves_vector                  Curves_vector;
-  
+
   typedef CGAL::Arrangement_on_surface_2<Geom_traits, Topol_traits>
                                                         Arrangement;
 
@@ -44,45 +44,45 @@ public:
   typedef typename Points_vector::iterator              Point_iterator;
   typedef std::vector<CGAL::Object>                     Objects_vector;
   typedef Objects_vector::iterator                      Object_iterator;
-  
+
 protected:
-  typedef typename CGAL::Arr_naive_point_location<Arrangement>     
+  typedef typename CGAL::Arr_naive_point_location<Arrangement>
                                                     Naive_point_location;
-  typedef typename CGAL::Arr_simple_point_location<Arrangement>     
+  typedef typename CGAL::Arr_simple_point_location<Arrangement>
                                                     Simple_point_location;
-  typedef typename CGAL::Arr_walk_along_line_point_location<Arrangement> 
+  typedef typename CGAL::Arr_walk_along_line_point_location<Arrangement>
                                                     Walk_point_location;
-  typedef typename CGAL::Arr_landmarks_point_location<Arrangement> 
+  typedef typename CGAL::Arr_landmarks_point_location<Arrangement>
                                                     Lm_point_location;
   typedef typename CGAL::Arr_random_landmarks_generator<Arrangement>
                                                     Random_lm_generator;
   typedef typename CGAL::Arr_landmarks_point_location<Arrangement,
-                                                      Random_lm_generator> 
+                                                      Random_lm_generator>
                                                     Lm_random_point_location;
   typedef typename CGAL::Arr_grid_landmarks_generator<Arrangement>
                                                     Grid_lm_generator;
   typedef typename CGAL::Arr_landmarks_point_location<Arrangement,
-                                                      Grid_lm_generator> 
+                                                      Grid_lm_generator>
                                                     Lm_grid_point_location;
   typedef typename CGAL::Arr_halton_landmarks_generator<Arrangement>
                                                     Halton_lm_generator;
   typedef typename CGAL::Arr_landmarks_point_location<Arrangement,
-                                                      Halton_lm_generator> 
+                                                      Halton_lm_generator>
                                                     Lm_halton_point_location;
   typedef typename CGAL::Arr_middle_edges_landmarks_generator<Arrangement>
                                                     Middle_edges_generator;
   typedef typename CGAL::Arr_landmarks_point_location<Arrangement,
-                                                      Middle_edges_generator> 
+                                                      Middle_edges_generator>
     Lm_middle_edges_point_location;
   typedef typename CGAL::Arr_landmarks_specified_points_generator<Arrangement>
     Specified_points_generator;
   typedef typename CGAL::Arr_landmarks_point_location<Arrangement,
-                                                      Specified_points_generator> 
+                                                      Specified_points_generator>
     Lm_specified_points_point_location;
-  typedef typename CGAL::Arr_trapezoid_ric_point_location<Arrangement> 
+  typedef typename CGAL::Arr_trapezoid_ric_point_location<Arrangement>
     Trapezoid_ric_point_location;
 
-  //   typedef CGAL::Arr_triangulation_point_location<Arrangement> 
+  //   typedef CGAL::Arr_triangulation_point_location<Arrangement>
   //     Triangulation_point_location;
 
   // ===> Add new point location type here <===
@@ -90,9 +90,9 @@ protected:
 protected:
   /*! The geometry traits */
   const Geom_traits& m_geom_traits;
-  
+
   /*! The arrangement */
-  Arrangement* m_arr;  
+  Arrangement* m_arr;
 
   /*! The input data file of the query points*/
   std::string m_filename_queries;
@@ -118,11 +118,11 @@ protected:
   Halton_lm_generator* m_halton_g;
   Middle_edges_generator* m_middle_edges_g;
   Specified_points_generator* m_specified_points_g;
-  
+
   // // ===> Change the number of point-location startegies
   // //      when a new point location is added. <===
   #define MAX_NUM_POINT_LOCATION_STRATEGIES 11
-  
+
 public:
   /*! Constructor */
   Point_location_test(const Geom_traits& geom_traits);
@@ -137,10 +137,10 @@ public:
 
   void set_filenames(const char* points_filename, const char* xcurves_filename,
                      const char* curves_filename, const char* queries_filename);
-  
+
   /*! Initialize the data structures */
   virtual bool init();
-  
+
   /*! Perform the test */
   virtual bool perform();
 
@@ -150,11 +150,11 @@ public:
   bool allocate_arrangement();
 
   void deallocate_arrangement();
-  
+
   bool construct_arrangement();
 
   void clear_arrangement();
-  
+
   bool allocate_pl_strategies();
 
   bool construct_pl_strategies();
@@ -162,14 +162,14 @@ public:
   bool attach_pl_strategies();
 
   void deallocate_pl_strategies();
-  
+
   template <typename Point_location,
             typename InputIterator, typename OutputIterator>
   void query(Point_location& pl, const char* type,
              InputIterator begin, InputIterator end, OutputIterator oi)
   {
     typedef InputIterator                       Input_iterator;
-    
+
     CGAL::Timer timer;
     timer.reset(); timer.start();
     Input_iterator piter;
@@ -184,7 +184,7 @@ public:
 };
 
 /*!
- * Constructor. 
+ * Constructor.
  */
 template <typename T_Geom_traits, typename T_Topol_traits>
 Point_location_test<T_Geom_traits, T_Topol_traits>::
@@ -192,22 +192,22 @@ Point_location_test(const Geom_traits& geom_traits) :
   Base(geom_traits),
   m_geom_traits(geom_traits),
   m_arr(NULL),
-  m_naive_pl(NULL), 
-  m_simple_pl(NULL), 
-  m_walk_pl(NULL), 
-  m_lm_pl(NULL), 
-  m_random_lm_pl(NULL), 
-  m_grid_lm_pl(NULL), 
-  m_halton_lm_pl(NULL), 
-  m_middle_edges_lm_pl(NULL), 
-  m_specified_points_lm_pl(NULL), 
-  // m_triangulation_pl(NULL), 
-  m_trapezoid_ric_pl(NULL), 
-  m_trapezoid_ric_no_grnt_pl(NULL), 
-  m_random_g(NULL), 
-  m_grid_g(NULL), 
-  m_halton_g(NULL), 
-  m_middle_edges_g(NULL), 
+  m_naive_pl(NULL),
+  m_simple_pl(NULL),
+  m_walk_pl(NULL),
+  m_lm_pl(NULL),
+  m_random_lm_pl(NULL),
+  m_grid_lm_pl(NULL),
+  m_halton_lm_pl(NULL),
+  m_middle_edges_lm_pl(NULL),
+  m_specified_points_lm_pl(NULL),
+  // m_triangulation_pl(NULL),
+  m_trapezoid_ric_pl(NULL),
+  m_trapezoid_ric_no_grnt_pl(NULL),
+  m_random_g(NULL),
+  m_grid_g(NULL),
+  m_halton_g(NULL),
+  m_middle_edges_g(NULL),
   m_specified_points_g(NULL)
 {}
 
@@ -350,10 +350,10 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::construct_arrangement()
   CGAL::insert(*m_arr, this->m_xcurves.begin(), this->m_xcurves.end());
   // insert(*m_arr, m_points.begin(), m_points.end());
   CGAL::insert(*m_arr, this->m_curves.begin(), this->m_curves.end());
-  
+
   // Print the size of the arrangement.
   std::cout << "V = " << m_arr->number_of_vertices()
-            << ",  E = " << m_arr->number_of_edges() 
+            << ",  E = " << m_arr->number_of_edges()
             << ",  F = " << m_arr->number_of_faces() << std::endl;
 
   return true;
@@ -389,7 +389,7 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::allocate_pl_strategies(
   //   return false;
 
 #endif
-  
+
   if (!(m_trapezoid_ric_pl = new Trapezoid_ric_point_location())) return false;
   if (!(m_trapezoid_ric_no_grnt_pl = new Trapezoid_ric_point_location()))
     return false;
@@ -416,32 +416,32 @@ construct_pl_strategies()
 
   timer.reset(); timer.start();
   m_lm_pl = new Lm_point_location(*m_arr);                              // 3
-  timer.stop(); 
+  timer.stop();
   std::cout << "Lm (vert) construction took " << timer.time() << std::endl;
 
   timer.reset(); timer.start();
   m_random_g = new Random_lm_generator(*m_arr);
   m_random_lm_pl = new Lm_random_point_location(*m_arr, m_random_g);    // 4
-  timer.stop(); 
+  timer.stop();
   std::cout << "Random lm construction took " << timer.time() << std::endl;
 
   timer.reset(); timer.start();
-  m_grid_g = new Grid_lm_generator(*m_arr);      
+  m_grid_g = new Grid_lm_generator(*m_arr);
   m_grid_lm_pl = new Lm_grid_point_location(*m_arr, m_grid_g);          // 5
-  timer.stop(); 
+  timer.stop();
   std::cout << "Grid lm construction took " << timer.time() << std::endl;
 
   timer.reset(); timer.start();
   m_halton_g = new Halton_lm_generator(*m_arr);
   m_halton_lm_pl = new Lm_halton_point_location(*m_arr, m_halton_g);    // 6
-  timer.stop(); 
+  timer.stop();
   std::cout << "Halton lm construction took " << timer.time() << std::endl;
 
   timer.reset(); timer.start();
   m_middle_edges_g = new Middle_edges_generator(*m_arr);
   m_middle_edges_lm_pl =
     new Lm_middle_edges_point_location(*m_arr, m_middle_edges_g);       // 7
-  timer.stop(); 
+  timer.stop();
   std::cout << "Middle edges lm construction took " << timer.time()
             << std::endl;
 
@@ -449,30 +449,30 @@ construct_pl_strategies()
   m_specified_points_g = new Specified_points_generator(*m_arr);
   m_specified_points_lm_pl =
     new Lm_specified_points_point_location(*m_arr, m_specified_points_g); // 8
-  timer.stop(); 
+  timer.stop();
   std::cout << "Specified_points lm construction took "
             << timer.time() << std::endl;
 
   // timer.reset(); timer.start();
   // m_triangulation_pl = new Triangulation_point_location(*m_arr);     // 9
-  // timer.stop(); 
+  // timer.stop();
   // std::cout << "Triangulation lm construction took "
   //           << timer.time() << std::endl;
 
 #endif
-  
+
   timer.reset(); timer.start();
   m_trapezoid_ric_pl = new Trapezoid_ric_point_location(*m_arr);        // 10
-  timer.stop(); 
+  timer.stop();
   std::cout << "Trapezoid RIC construction took " << timer.time() << std::endl;
-  
+
   timer.reset(); timer.start();
   m_trapezoid_ric_no_grnt_pl =
     new Trapezoid_ric_point_location(*m_arr, false);                    // 11
   timer.stop();
-  std::cout << "Trapezoid RIC without-guarantees construction took " 
+  std::cout << "Trapezoid RIC without-guarantees construction took "
             << timer.time() << std::endl;
-  
+
   std::cout << std::endl;
 
   // ===> Add new point location instance here. <===
@@ -496,31 +496,31 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::attach_pl_strategies()
 
   timer.reset(); timer.start();
   m_lm_pl->attach(*m_arr);
-  timer.stop(); 
+  timer.stop();
   std::cout << "Lm (vert) construction took " << timer.time() << std::endl;
 
   timer.reset(); timer.start();
   m_random_g = new Random_lm_generator(*m_arr);
   m_random_lm_pl->attach(*m_arr, m_random_g);
-  timer.stop(); 
+  timer.stop();
   std::cout << "Random lm construction took " << timer.time() << std::endl;
 
   timer.reset(); timer.start();
-  m_grid_g = new Grid_lm_generator(*m_arr);      
+  m_grid_g = new Grid_lm_generator(*m_arr);
   m_grid_lm_pl->attach(*m_arr, m_grid_g);
-  timer.stop(); 
+  timer.stop();
   std::cout << "Grid lm construction took " << timer.time() << std::endl;
 
   timer.reset(); timer.start();
   m_halton_g = new Halton_lm_generator(*m_arr);
   m_halton_lm_pl->attach(*m_arr, m_halton_g);
-  timer.stop(); 
+  timer.stop();
   std::cout << "Halton lm construction took " << timer.time() << std::endl;
 
   timer.reset(); timer.start();
   m_middle_edges_g = new Middle_edges_generator(*m_arr);
   m_middle_edges_lm_pl->attach(*m_arr, m_middle_edges_g);
-  timer.stop(); 
+  timer.stop();
   std::cout << "Middle edges lm construction took " << timer.time()
             << std::endl;
 
@@ -533,25 +533,25 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::attach_pl_strategies()
 
   // timer.reset(); timer.start();
   // m_location triangulation_lm_pl->attach(*m_arr);
-  // timer.stop(); 
+  // timer.stop();
   // std::cout << "Triangulation lm construction took "
   //           << timer.time() << std::endl;
 
 #endif
-  
+
   timer.reset(); timer.start();
   m_trapezoid_ric_pl->attach(*m_arr);
-  timer.stop(); 
+  timer.stop();
   std::cout << "Trapezoid RIC construction took " << timer.time() << std::endl;
-  
+
   timer.reset(); timer.start();
   m_trapezoid_ric_no_grnt_pl->with_guarantees(false);
   m_trapezoid_ric_no_grnt_pl->attach(*m_arr);
 
   timer.stop();
-  std::cout << "Trapezoid RIC without-guarantees construction took " 
+  std::cout << "Trapezoid RIC without-guarantees construction took "
             << timer.time() << std::endl;
-  
+
   std::cout << std::endl;
 
   // ===> Add new point location instance here. <===
@@ -566,7 +566,7 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::perform()
   typename Arrangement::Vertex_const_handle    vh_ref, vh_curr;
   typename Arrangement::Halfedge_const_handle  hh_ref, hh_curr;
   typename Arrangement::Face_const_handle      fh_ref, fh_curr;
-  
+
   // Locate the points in the list using all point location strategies.
 
   // std::cout << "Time in seconds" << std::endl;
@@ -579,7 +579,7 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::perform()
 
   query(*m_simple_pl, "Simple", m_query_points.begin(), m_query_points.end(),
         std::back_inserter(objs[pl_index++]));  // Simple
-  
+
   query(*m_walk_pl, "Walk", m_query_points.begin(), m_query_points.end(),
         std::back_inserter(objs[pl_index++]));  // Walk
 
@@ -609,7 +609,7 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::perform()
   query(*m_specified_points_lm_pl, "Landmarks specified points",
         m_query_points.begin(), m_query_points.end(),
         std::back_inserter(objs[pl_index++]));  // Landmarks specified points
-  
+
   // Triangulation
   // query(*m_triangulation_pl, "Triangulation",
   //       m_query_points.begin(), m_query_points.end(),
@@ -632,7 +632,7 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::perform()
 
   // Number of point location strategies used.
   unsigned int pls_num = pl_index;
-  std::cout << "Number of strategies is " << pls_num << std::endl;  
+  std::cout << "Number of strategies is " << pls_num << std::endl;
 
   // End Location
 
@@ -644,7 +644,7 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::perform()
     ob_iter[pl_index] = objs[pl_index].begin();
 
   // get size of objects
-  unsigned int size = static_cast<unsigned int>(objs[0].size());
+  size_t size = objs[0].size();
   std::cout << "size is " << size << std::endl;
 
   for (pl_index = 0; pl_index < pls_num; ++pl_index) {
@@ -678,11 +678,11 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::perform()
           result = -1;
         }
         else if (fh_curr != fh_ref) {
-          std::cout << "Error: point location number " 
+          std::cout << "Error: point location number "
                     << pl_index << " return a different face" << std::endl;
           result = -1;
         }
-      }  
+      }
       //if (fh_ref->is_unbounded())
       //  std::cout << "Unbounded face." << std::endl;
       //else
@@ -710,7 +710,7 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::perform()
           result = -1;
         }
         else if ((hh_curr != hh_ref) && (hh_curr->twin() != hh_ref)) {
-          std::cout << "Error: point location number " 
+          std::cout << "Error: point location number "
                     << pl_index << " return a different halfedge" << std::endl;
           std::cout << "Halfedge (curr): " << hh_curr->curve() << std::endl;
           result = -1;
@@ -737,7 +737,7 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::perform()
           result = -1;
         }
         else if (vh_curr != vh_ref) {
-          std::cout << "Error: point location number " 
+          std::cout << "Error: point location number "
                     << pl_index << " return a different vertex"<< std::endl;
           result = -1;
         }
@@ -746,7 +746,7 @@ bool Point_location_test<T_Geom_traits, T_Topol_traits>::perform()
     }
 
     else {
-      std::cout << "Illegal point-location result." << std::endl;    
+      std::cout << "Illegal point-location result." << std::endl;
       result = -1;
     }
   }
